Moved this test case to test_nodes_maintenance.py file and added DC pods
