+++ tests/manage/z_cluster/nodes/test_nodes_maintenance.py
@@ -3,12 +3,16 @@ import pytest
 
 from subprocess import TimeoutExpired
 
-from ocs_ci.ocs.exceptions import CephHealthException
+from ocs_ci.ocs.exceptions import (
+    CephHealthException, ResourceWrongStatusException
+)
 from ocs_ci.utility.utils import ceph_health_check_base
 
-from ocs_ci.ocs import constants
+from ocs_ci.ocs import constants, machine, ocp
 from ocs_ci.ocs.node import (
-    drain_nodes, schedule_nodes, get_typed_nodes, wait_for_nodes_status, get_node_objs
+    drain_nodes, schedule_nodes, get_typed_nodes, wait_for_nodes_status,
+    remove_nodes, get_osd_running_nodes, get_node_objs,
+    add_new_node_and_label_it
 )
 from ocs_ci.framework.testlib import (
     tier1, tier2, tier3, tier4, tier4b,
@@ -16,18 +20,25 @@ from ocs_ci.framework.testlib import (
 )
 
 from tests.sanity_helpers import Sanity
+from ocs_ci.ocs.resources import pod
+from tests.helpers import (
+    label_worker_node, remove_label_from_worker_node)
+from tests import helpers
 
-logger = logging.getLogger(__name__)
+
+log = logging.getLogger(__name__)
 
 
 @pytest.fixture(autouse=True)
-def schedule_nodes_teardown(request):
+def teardown(request):
     """
-    Make sure that all cluster's nodes are in 'Ready' state and if not,
-    change them back to 'Ready' state by marking them as scheduble
+
+    Tear down function
 
     """
     def finalizer():
+        # Make sure that all cluster's nodes are in 'Ready' state and if not,
+        # change them back to 'Ready' state by marking them as schedulable
         scheduling_disabled_nodes = [
             n.name for n in get_node_objs() if n.ocp.get_resource_status(
                 n.name
@@ -35,6 +46,14 @@ def schedule_nodes_teardown(request):
         ]
         if scheduling_disabled_nodes:
             schedule_nodes(scheduling_disabled_nodes)
+
+        # Remove label created for DC app pods on all worker nodes
+        node_objs = get_node_objs()
+        for node_obj in node_objs:
+            if 'dc' in node_obj.get().get('metadata').get('labels').keys():
+                remove_label_from_worker_node(
+                    [node_obj.name], label_key="dc"
+                )
     request.addfinalizer(finalizer)
 
 
@@ -63,7 +82,7 @@ class TestNodesMaintenance(ManageTest):
         try:
             status = ceph_health_check_base()
             if status:
-                logger.info("Health check passed")
+                log.info("Health check passed")
         except CephHealthException as e:
             # skip because ceph is not in good health
             pytest.skip(str(e))
@@ -178,7 +197,7 @@ class TestNodesMaintenance(ManageTest):
         try:
             drain_nodes(typed_node_names)
         except TimeoutExpired:
-            logger.info(f"Draining of nodes {typed_node_names} failed as expected")
+            log.info(f"Draining of nodes {typed_node_names} failed as expected")
 
         schedule_nodes(typed_node_names)
 
@@ -219,6 +238,126 @@ class TestNodesMaintenance(ManageTest):
 
         # Mark the nodes back to schedulable
         schedule_nodes(node_names)
+
+        # Perform cluster and Ceph health checks
+        self.sanity_helpers.health_check()
+
+    @tier4
+    @tier4b
+    @aws_platform_required
+    @pytest.mark.parametrize(
+        argnames=["interface"],
+        argvalues=[
+            pytest.param(
+                *['rbd'],
+                marks=pytest.mark.polarion_id("OCS-2128")
+            ),
+            pytest.param(
+                *['cephfs'],
+                marks=pytest.mark.polarion_id("OCS-2129")
+            ),
+        ]
+    )
+    def test_simultaneous_drain_of_two_ocs_nodes(
+        self, pvc_factory, pod_factory, dc_pod_factory,
+        interface
+    ):
+        """
+
+        Simultaneous drain of two OCS nodes
+
+        """
+        # Get OSD running nodes
+        osd_running_worker_nodes = get_osd_running_nodes()
+        log.info(f"OSDs are running on nodes {osd_running_worker_nodes}")
+
+        # Label osd nodes with fedora app
+        label_worker_node(
+            osd_running_worker_nodes, label_key='dc', label_value='fedora'
+        )
+        log.info("Successfully labeled worker nodes with {dc:fedora}")
+
+        # Create DC app pods
+        log.info("Creating DC based app pods and starting IO in background")
+        if interface == 'rbd':
+            interface = constants.CEPHBLOCKPOOL
+        elif interface == 'cephfs':
+            interface = constants.CEPHFILESYSTEM
+        dc_pod_obj = []
+        for i in range(2):
+            dc_pod = dc_pod_factory(
+                interface=interface, node_selector={'dc': 'fedora'}
+            )
+            pod.run_io_in_bg(dc_pod, fedora_dc=True)
+            dc_pod_obj.append(dc_pod)
+
+        # Get the machine name using the node name
+        machine_names = [
+            machine.get_machine_from_node_name(osd_running_worker_node)
+            for osd_running_worker_node in osd_running_worker_nodes[:2]
+        ]
+        log.info(f"{osd_running_worker_nodes} associated "
+                 f"machine are {machine_names}")
+
+        # Get the machineset name using machine name
+        machineset_names = [
+            machine.get_machineset_from_machine_name(
+                machine_name
+            )
+            for machine_name in machine_names
+        ]
+        log.info(
+            f"{osd_running_worker_nodes} associated machineset "
+            f"is {machineset_names}"
+        )
+
+        # Add a new node and label it
+        add_new_node_and_label_it(machineset_names[0])
+        add_new_node_and_label_it(machineset_names[1])
+
+        # Drain 2 nodes
+        drain_nodes(osd_running_worker_nodes[:2])
+
+        # Check the pods should be in running state
+        all_pod_obj = pod.get_all_pods(wait=True)
+        for pod_obj in all_pod_obj:
+            if '-1-deploy' and 'ocs-deviceset' not in pod_obj.name:
+                try:
+                    helpers.wait_for_resource_state(
+                        resource=pod_obj, state=constants.STATUS_RUNNING,
+                        timeout=200
+                    )
+                except ResourceWrongStatusException:
+                    # 'rook-ceph-crashcollector' on the failed node stucks at
+                    # pending state. BZ 1810014 tracks it.
+                    # Ignoring 'rook-ceph-crashcollector' pod health check as
+                    # WA and deleting its deployment so that the pod
+                    # disappears. Will revert this WA once the BZ is fixed
+                    if 'rook-ceph-crashcollector' in pod_obj.name:
+                        ocp_obj = ocp.OCP()
+                        name = pod_obj.name[:-17]
+                        command = f"delete deployment {name}"
+                        ocp_obj.exec_oc_cmd(command=command)
+                        log.info(f"Deleted deployment for pod {pod_obj.name}")
+
+        # DC app pods on the drained node will get automatically created on other
+        # running node in same AZ. Waiting for all dc app pod to reach running state
+        pod.wait_for_dc_app_pods_to_reach_running_state(dc_pod_obj)
+        log.info("All the dc pods reached running state")
+
+        # Remove unscheduled nodes
+        # In scenarios where the drain is attempted on >3 worker setup,
+        # post completion of drain we are removing the unscheduled nodes so
+        # that we maintain 3 worker nodes.
+        log.info(f"Removing scheduled nodes {osd_running_worker_nodes[:2]}")
+        remove_node_objs = get_node_objs(osd_running_worker_nodes[:2])
+        remove_nodes(remove_node_objs)
+
+        # Check basic cluster functionality by creating resources
+        # (pools, storageclasses, PVCs, pods - both CephFS and RBD),
+        # run IO and delete the resources
+        self.sanity_helpers.create_resources(pvc_factory, pod_factory)
+        self.sanity_helpers.delete_resources()
 
         # Perform cluster and Ceph health checks
         self.sanity_helpers.health_check()
